Move the pipeline graph manipulation utilites to their own files.
As well as being very different from the low-level execution details
found in common.py, there is no need to compile these (whereas the
other structures in common.py are critical to optimize for elementwise
efficiency).
