+++ sdks/python/apache_beam/runners/portability/fn_api_runner/fn_runner.py
@@ -62,10 +62,10 @@ from apache_beam.portability.api import beam_provision_api_pb2
 from apache_beam.portability.api import beam_runner_api_pb2
 from apache_beam.portability.api import metrics_pb2
 from apache_beam.runners import runner
-from apache_beam.runners.common import group_by_key_input_visitor
-from apache_beam.runners.common import merge_common_environments
-from apache_beam.runners.common import merge_superset_dep_environments
-from apache_beam.runners.common import validate_pipeline_graph
+from apache_beam.runners.pipeline_utils import group_by_key_input_visitor
+from apache_beam.runners.pipeline_utils import merge_common_environments
+from apache_beam.runners.pipeline_utils import merge_superset_dep_environments
+from apache_beam.runners.pipeline_utils import validate_pipeline_graph
 from apache_beam.runners.portability import portable_metrics
 from apache_beam.runners.portability.fn_api_runner import execution
 from apache_beam.runners.portability.fn_api_runner import translations
@@ -218,7 +218,7 @@ class FnApiRunner(runner.PipelineRunner):
       pipeline_proto = self.embed_default_docker_image(pipeline_proto)
     pipeline_proto = merge_common_environments(
         self.resolve_any_environments(
-          merge_superset_dep_environments(pipeline_proto)))
+            merge_superset_dep_environments(pipeline_proto)))
     stage_context, stages = self.create_stages(pipeline_proto)
     return self.run_stages(stage_context, stages)
